@@ -1,7 +1,5 @@
 """Retain apiclient as an alias for googleapiclient."""
 
-from six import iteritems
-
 import googleapiclient
 
 from googleapiclient import channel
@@ -19,8 +17,6 @@ except ImportError:
     sample_tools = None
 from googleapiclient import schema
 
-__version__ = googleapiclient.__version__
-
 _SUBMODULES = {
     "channel": channel,
     "discovery": discovery,
@@ -34,5 +30,5 @@ _SUBMODULES = {
 
 import sys
 
-for module_name, module in iteritems(_SUBMODULES):
+for module_name, module in _SUBMODULES.items():
     sys.modules["apiclient.%s" % module_name] = module
