@@ -76,7 +76,6 @@ import uuid
 
 from googleapiclient import errors
 from googleapiclient import _helpers as util
-import six
 
 
 # The unix time epoch starts at midnight 1970.
@@ -104,7 +103,7 @@ X_GOOG_RESOURCE_ID = "X-GOOG-RESOURCE-ID"
 
 def _upper_header_keys(headers):
     new_headers = {}
-    for k, v in six.iteritems(headers):
+    for k, v in headers.items():
         new_headers[k.upper()] = v
     return new_headers
 
@@ -244,7 +243,7 @@ class Channel(object):
     Args:
       resp: dict, The response from a watch() method.
     """
-        for json_name, param_name in six.iteritems(CHANNEL_PARAMS):
+        for json_name, param_name in CHANNEL_PARAMS.items():
             value = resp.get(json_name)
             if value is not None:
                 setattr(self, param_name, value)
