@@ -43,46 +43,56 @@ class HttpError(Error):
         self.content = content
         self.uri = uri
         self.error_details = ""
+        self.reason = self._get_reason()
+
+    @property
+    def status_code(self):
+        """Return the HTTP status code from the response content."""
+        return self.resp.status
 
     def _get_reason(self):
         """Calculate the reason for the error from the response content."""
         reason = self.resp.reason
         try:
-            data = json.loads(self.content.decode("utf-8"))
+            try:
+                data = json.loads(self.content.decode("utf-8"))
+            except json.JSONDecodeError:
+                # In case it is not json
+                data = self.content.decode("utf-8")
             if isinstance(data, dict):
                 reason = data["error"]["message"]
-                if "details" in data["error"]:
-                    self.error_details = data["error"]["details"]
-                elif "detail" in data["error"]:
-                    self.error_details = data["error"]["detail"]
+                error_detail_keyword = next((kw for kw in ["detail", "details", "errors", "message"] if kw in data["error"]), "")
+                if error_detail_keyword:
+                    self.error_details = data["error"][error_detail_keyword]
             elif isinstance(data, list) and len(data) > 0:
                 first_error = data[0]
                 reason = first_error["error"]["message"]
                 if "details" in first_error["error"]:
                     self.error_details = first_error["error"]["details"]
+            else:
+                self.error_details = data
         except (ValueError, KeyError, TypeError):
             pass
         if reason is None:
             reason = ""
-        return reason
+        return reason.strip()
 
     def __repr__(self):
-        reason = self._get_reason()
         if self.error_details:
             return '<HttpError %s when requesting %s returned "%s". Details: "%s">' % (
                 self.resp.status,
                 self.uri,
-                reason.strip(),
+                self.reason,
                 self.error_details,
             )
         elif self.uri:
             return '<HttpError %s when requesting %s returned "%s">' % (
                 self.resp.status,
                 self.uri,
-                self._get_reason().strip(),
+                self.reason,
             )
         else:
-            return '<HttpError %s "%s">' % (self.resp.status, self._get_reason())
+            return '<HttpError %s "%s">' % (self.resp.status, self.reason)
 
     __str__ = __repr__
 
@@ -124,7 +134,7 @@ class MediaUploadSizeError(Error):
 
 
 class ResumableUploadError(HttpError):
-    """Error occured during resumable upload."""
+    """Error occurred during resumable upload."""
 
     pass
 
@@ -142,7 +152,7 @@ class InvalidNotificationError(Error):
 
 
 class BatchError(HttpError):
-    """Error occured during batch operations."""
+    """Error occurred during batch operations."""
 
     @util.positional(2)
     def __init__(self, reason, resp=None, content=None):
