@@ -3,8 +3,8 @@ from __future__ import absolute_import
 
 __author__ = "jcgregorio@google.com (Joe Gregorio)"
 
-import unittest2 as unittest
 import datetime
+import unittest
 
 from googleapiclient import channel
 from googleapiclient import errors
