@@ -22,14 +22,10 @@ from __future__ import absolute_import
 
 __author__ = "mmcdonald@google.com (Matt McDonald)"
 
-import unittest2 as unittest
+import unittest
 import httplib2
-import googleapiclient.model
 
-from googleapiclient.errors import HttpError
 from googleapiclient.model import ProtocolBufferModel
-
-from six.moves.urllib.parse import parse_qs
 
 
 class MockProtocolBuffer(object):
