@@ -19,7 +19,7 @@ __author__ = "jcgregorio@google.com (Joe Gregorio)"
 
 import json
 import os
-import unittest2 as unittest
+import unittest
 
 from googleapiclient.schema import Schemas
 
@@ -32,20 +32,20 @@ def datafile(filename):
 
 
 LOAD_FEED = """{
-    "items": [
-      {
-        "longVal": 42,
-        "kind": "zoo#loadValue",
-        "enumVal": "A String",
-        "anyVal": "", # Anything will do.
-        "nullVal": None,
-        "stringVal": "A String",
-        "doubleVal": 3.14,
-        "booleanVal": True or False, # True or False.
-      },
-    ],
-    "kind": "zoo#loadFeed",
-  }"""
+  "items": [
+    {
+      "longVal": 42,
+      "kind": "zoo#loadValue",
+      "enumVal": "A String",
+      "anyVal": "", # Anything will do.
+      "nullVal": None,
+      "stringVal": "A String",
+      "doubleVal": 3.14,
+      "booleanVal": True or False, # True or False.
+    },
+  ],
+  "kind": "zoo#loadFeed",
+}"""
 
 
 class SchemasTest(unittest.TestCase):
